requirements.txt
django
gunicorn
django-heroku
line-bot-sdk
